+++ index.js
@@ -11,6 +11,7 @@ module.exports = {
     'jsx-prefer-fragment-wrappers': require('./lib/rules/jsx-prefer-fragment-wrappers'),
     'no-ancestor-directory-import': require('./lib/rules/no-ancestor-directory-import'),
     'no-debugger': require('./lib/rules/no-debugger'),
+    'no-namespace-imports': require('./lib/rules/no-namespace-imports'),
     'no-useless-computed-properties': require('./lib/rules/no-useless-computed-properties'),
     'no-fully-static-classes': require('./lib/rules/no-fully-static-classes'),
     'polaris-prefer-sectioned-prop': require('./lib/rules/polaris-prefer-sectioned-prop'),