+++ CHANGELOG.md
@@ -2,13 +2,16 @@
 
 ## Unreleased
 
+### New Rules
+
 - `shopify/no-all-mocks-methods` ([#204](https://github.com/Shopify/eslint-plugin-shopify/pull/204))
-- `jest/valid-title`
-- `jest/prefer-hooks-on-top`
-- `jest/require-top-level-describe`
+- `shopify/no-namespace-imports` Prevent namespace import declarations. ([262](https://github.com/Shopify/eslint-plugin-shopify/pull/262))
 
 ## [32.0.0] - 2019-11-05
 
+- `jest/valid-title`
+- `jest/prefer-hooks-on-top`
+- `jest/require-top-level-describe`
 - Enforce new-lines between groups import groups ([#409](https://github.com/Shopify/eslint-plugin-shopify/pull/409))
 
 ## [31.0.0] - 2019-10-23